requirements.txt
websockets==12.0
python-dotenv==1.0.0
screeninfo==0.8.1